+++ tfjs-core/src/util.ts
@@ -405,18 +405,8 @@ export function getArrayFromDType<D extends DataType>(
   return values as DataTypeMap[D];
 }
 
-export function checkComputationForErrors<D extends DataType>(
-    vals: DataTypeMap[D], dtype: D, name: string): void {
-  if (dtype !== 'float32') {
-    // Only floating point computations will generate NaN values
-    return;
-  }
-  for (let i = 0; i < vals.length; i++) {
-    const num = vals[i] as number;
-    if (isNaN(num) || !isFinite(num)) {
-      throw Error(`The result of the '${name}' is ${num}.`);
-    }
-  }
+export function UserException(msg: string) {
+  console.error(msg);
 }
 
 export function checkConversionForErrors<D extends DataType>(
